Move settings into configuration namespace
