+++ ABSoftware.ABSave/Configuration/SettingsBuilder.cs
@@ -1,20 +1,10 @@
-﻿using ABCo.ABSave.Converters;
-using ABCo.ABSave.Helpers;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Reflection;
 using System.Text;
 
-namespace ABCo.ABSave
+namespace ABCo.ABSave.Configuration
 {
-    public enum SettingsPreset
-    {
-        PrioritizePerformance,
-        PrioritizeSize
-    }
-
-    public struct ABSaveSettingsBuilder
+    public struct SettingsBuilder
     {
         public bool? LazyBitHandling { get; set; }
         public bool? UseUTF8 { get; set; }
@@ -60,38 +50,5 @@ namespace ABCo.ABSave
             return new ABSaveSettings(lazyBitHandling, useUTF8, bypassDangerousTypeChecking, useLittleEndian,
                 exactConverters ?? Converter.BuiltInExact, nonExactConverters ?? Converter.BuiltInNonExact);
         }
-    }
-
-    public enum ABSavePresets
-    {
-        SpeedFocusInheritance,
-        SpeedFocusNoInheritance,
-        SizeFocusInheritance,
-        SizeFocusNoInheritance
-    }
-
-    /// <summary>
-    /// Stores the configuration for serialization/deserialization.
-    /// </summary>
-    public class ABSaveSettings
-    {
-        public static ABSaveSettings ForSpeed { get; } = new ABSaveSettings(true, true, false, true, Converter.BuiltInExact, Converter.BuiltInNonExact);
-        public static ABSaveSettings ForSize { get; } = new ABSaveSettings(false, true, false, true, Converter.BuiltInExact, Converter.BuiltInNonExact);
-
-        public bool LazyBitHandling { get; } = true;
-        public bool UseUTF8 { get; } = true;
-        public bool BypassDangerousTypeChecking { get; set; } = false;
-        public bool UseLittleEndian { get; } = true;
-
-        public IReadOnlyDictionary<Type, Converter>? ExactConverters { get; } = Converter.BuiltInExact;
-        public IReadOnlyList<Converter>? NonExactConverters { get; } = Converter.BuiltInNonExact;
-
-        public ABSaveSettings() { }
-
-        public ABSaveSettings(bool lazyBitHandling, bool useUTF8, bool bypassDangerousTypeChecking, bool useLittleEndian,
-            IReadOnlyDictionary<Type, Converter>? exactConverters, IReadOnlyList<Converter>? nonExactConverters)
-        =>
-            (LazyBitHandling, UseUTF8, UseLittleEndian, BypassDangerousTypeChecking, ExactConverters, NonExactConverters) = 
-            (lazyBitHandling, useUTF8, useLittleEndian, bypassDangerousTypeChecking, exactConverters, nonExactConverters);
     }
 }
